Add static version of file removal function in FiledropAccount
Normally the directories associated with a file drop account are removed
when an instance of the PHP class FiledropAccount is updated with a terminated status.

There's already a method ``removeDirectories()`` used for that.

However there are times when we need to remove files without a PHP instance existing.
so we need to have a static method for deletion we can call on the FiledropAccount class
directly.

I also correct the functions name to be more accurate.
The method is now called ``removeAccountDirectories()`` and doesn't require a parameter for
the identifier as it can access it through instance properties.
The static function is called ```removeDirectories()`` and take a parameter for the identifier.

Also, to avoid code duplication the method will call the static function.

Finally, we've modified the housekeeping function to remove archive filedrop account,
to use the static function to remove files if it cannot find a record in the database
(this is for situation were the table record was deleted but something went wrong before the files got removed.)