Add npm install command into scripts
npm install will download node package dependencies into a new
node_modules directory in the scripts folder if it is not present.
If node_modules is present then npm install will only audit these
packages for problems.

If there is no node_modules directory and either of the 2 setup
scripts are run then errors appear saying that node modules could
not be found.