Fix temporarily class map problem with ftp console controller
currently unable to fix  the build stage on gitlab ci
as the console service fail when installing composer dependency.

temnporary fix is to use magic number and remove dependency on the class for which
a PSR 4 entry in the composer config was necessary.

We cannot keep a magic number here, so later we need to look at the problem again.