Use internal id for looking up filedrop account and correct uploads
in upload table for FUW, doi is not unique as drop box account for the same doi
can be recreated several time, so when looking up for an upload associated to a file drop account,
internally we use the id for file drop account for the lookup.
I've already made a change in other area, this commit is for area I missed.

For the unit testing, the strategy is to add a dummy  upload with same DOI and file name but different filedrop account id