Fix error when connection cannot be established to the DOI API
Recently StoredDatasetConnections was updated to catch server exceptions that occur often
due to the doi.org API timeout out on requests, resulting in a Service Unavailable error ????

However there is another kind of failure that I have witnessed happening intermittently on my local dev environment
and on CI: The Guzzle client sends request to the doi.org API, but the latter queues the request in a waiting loop.
The "connect_timeout" parameter to Guzzle client actually determines how long it should maintain itself in the awaiting connection stage.
Intermittently (actually 50% of the time on my machine), that threshold will be triggered, causing an error that could not be caught by the Guzzle ServerException because connection wasn't established yet therefore there was no server response.

Additionally, whichever cause makes the API call to fail, it will result in the citation for the associated link to not be retrieved.
The test data has 3 links, and any of them could fail. On my development machine, the frequency of failure of at least two of the link
varies form intermittent to every other run, and on my CI pipeline, it's every time (does that mean that API call on CI is failing every time for some reasons?)

Anyway, the fact that the DatasetViewTest::testItShoulddisplayCitations() test tries to test the presence of the data contained in the API response is problematic, as it is not always there.

The fix is a two throng approach:

1. Add an other catch block to capture Guzzle ConnectException in StoredDatasetConnections.php
2. In the DatasetViewTest::testItShoulddisplayCitations(), assert the existence of the PubMed reference only, not the citation text

That is enough to assert that we've displayed all the information we can instead of user visible error messages on the dataset view page in the event of connection issues to the DOI API, which is the original intention.