Fix value of connect_timeout
It was changed to 30 seconds which is too long to wait as this is an API call within a user request,
we should return quicker to the user even if the API call is not successful, especially as there still data to show.

Also the unit test expected the original 5s still.

Now reverted to 5s.