Add Dockerfile and Docker compose for the updater command and a test DB
The Dockerfile, based on PHP 7.4 for CLI on Debian Buster will be the PHP environment under which
the Yii2 command will run. Also for running the acceptance tests.

Installation of client libraries for legacy PostgreSQL is also taken care of in the Dockerfile.

The compose YAML file is to spawn a daemon for a local legacy postgreSQL server to test the command locally
and for a convenient execution of the Yii2 command (take care of volumes)

An empty gigadb database and user is created when the database container service is started to facilitate restoring the backup file.