Configure the Yii2 project to function for CLI tool and be testable
Functional tests.
Inspired by what's done on FUW and from:
https://forum.yiiframework.com/t/tests-for-console-commands/86482/3

but both use Yii2 advanced template; for the basic template we need to modify ``console.php``
to find commands in the controllers directory instead of the commands directory so we can apply the same technique.

Also configured the functional and acceptance test suite with necessary modules.

Commited the composer.json and lock files for the project, that were missed the first time because of the gigadb level .gitignore
was blocking them. This is now  overriden in the local .gitignore.