Change setup script to make it easy to switch between test data CSV sets
when using setup_devdb.sh to switch between test data CSV sets, it may leads into
inconsistent state for the applications (esp. File Upload Wizard), because
the database migrations steps that were run at setup and steps to generate data feeds
from database input are not performed again.

Instead, the setup scrip up.sh, has now the ability to take as parameter the label of a set of test data
(with a default one if no parameter is supplied) and not only update the schema and tables appropriately
but also perform all the setup steps that depends on the database content.

Updated the README to describe those possibilities.