Implement default path test scenario for gigascience/gigadb-website#617
As part of task gigascience/gigadb-website#631, initial implementation of the test scenarios: it's not optimised and the tests are failing (for the right reason) because the functionality
is yet to be implemented.

The only thing the command does at the moment is the auxiliary prerequisite command to download the production backup and load it into a local PostgreSQL server running as container service.