Use Yii2 DB config to configure Db access needed by system calls
That way the database configuration is in one place (Yii2 config) whether we pass them to
psql through system() or when using Yii2 ActiveRecored, Data Access commands or Query buiilder script.

Modified the download-restore-backup command to use the Yii2 configuration.
More importantly that approach will allow writing functional tests that use the default Yii2 style of using
a test database.

to help with bootstraping the test database, there is now a sql database schema for Gigadb v3. It is called in the setup
method of the functional tests.

Corrected the functional test for download-restore-backup to use backup from the day before as the parrot ftp server don't seem to keep backup for
more than a week.