Setup unit tests and fixture data, add logic to list pending datasets
In order to unit test the business logic of ftp urls and ftp_site replacement,
I need to setup the Codeception unit test suite, the main config being ``unit.suite.yml``
which needs to be gitignored as it will have DSN in it. To use as a template there is
``unit.suite.yml.example``

``sql/gigadb_tables.sql`` will be used to bootstrap the test database,
but before we can use it,  we need to remove any mention of EXTENSION from the file
as they would require superuser to load them, and we want to use our default
user for that.
That user needs to have CREATEDB privilege in order to drop and create database, that's done in
``boostrap_gigadb.sql``. Also removed setting up password in that file.

As part of configuring the Codeception DB module in unit.suite.yml, we need a reset script
that will recreate the test database. That's what the ``repopulate_testdb.sh`` is for.

Created unit test case for our ``DatasetFiles.php``'s business logic in ``tests/units/models/DatasetFilesTest.php``

with fixture data created in the ``_before()`` setup method, using Codeception Db methods

In DatasetFiles.php model class, we've laid down, using Yii DB Query class, the business logic for
listing pending datasets in need of ftp urls replacement.