Add command to list pending datasets and use non-superuser for DB reset
with command line options designed for batch operations, so we can list
the next batch of specifiable size or all. For safety, not passing either options will do nothing.
The result is made using of the Yii2 DB Query queries from model class ``DatasetFiles.php``.

that list command is meant to morph into the dry-run mode of the ftp url replacement commands later.

The command to download production database and load it in local database was updated so it doesn't require
PostgreSQL superuser can be performed with default user. I've update the README with explanations on the gotcha.

Some basic functional tests are added to support the various options for listing pending datasets.