Add parameter to dataset listing command to allow a starting point
In addition of --next, can pass a --after parameter to the command listing pending datasets
so that we can return all the dataset of the batch size that are after that dataset_id position in the list.

Upddate the Yii DB query logic in the model class to handle that parameter too.

Unit tests were amended accordingly.

I also removed duplication in the business logic so that the part of the query common to all calling mode is
only specified once.