Add unit tests for other ftp_site hosts and support for no-op.
Added a couple of fixture data to allow that.
If ftp_site value already start with https, we do nothing.

Tested that the replacement work with  ftp://parrot.genomics.cn and ftp://climb.genomics.cn

Made the code more robust, so it always return the value of ftp_site if success or no-op
and log an error for every other cases.