Fix initialisation of parameters
To make the program robust, we declare the parameters as typed variables, but
that means they need to be initialised.

Made sure the initial value for $all and $next are falsy, so we can still return ExitCode::NOINPUT
if neither of those parameters are passed.

Also initialised the $date parameter used by the command for downloading and loading production backup to the day before yesterday,
so that that parameter is not required if one wants to retrieve the most recent backup of the production database.