Enable dry run mode in model class for ftp url replacement tool
In previous commit, we've segregated the database interaction into its own function,
so that the change to enable dry run mode can be localised there.
(because dry-run means we do everything for real except for the final update of database records)

Enabled a property and a constant on the model class that will allow the controller to translate the
CLI options --dry-run into actual activation of the mode.

Made all the database methods private as they should not be used outside the class.

Replaced ``->one()['ftp_site']`` by the proper form ``->scalar()``