Implement command for replacing ftp urls in location and ftp_site fields
Before, had to update the business logic in the model class, so that the method to return the files
for pending datasets only return files with location that need replacing.

Then implement in the controller call to the model method for making the replacement of ftp_site
and location column.

some work had to be done to translate the options passed by the users into parameters to pass on
to the replacement method for the batch size, starting point, and activation of dry-run mode.

The main output of the command lists the dataset being processed and a three colour coded flag to quickly express level
of success, as well as the number of file processed compared to the total number of files in need of replacement.

Also added a --verbose option to allow displaying the audit table generated by the replacement methods.