When recreating database, terminate DB connections beforehand
the command for downloading production database backup and loading it in local server
occasionally fail with Postgres error about other users being already connected,
thus preventing the database to be dropped and re-created.

Fixed it by having a query to terminate all db connections to precede the drop/create instructions.