Add no download option to command to fetch and reload production backup
Because we want to use the same command in _before hook for acceptance tests
using a stored production backup, so we don't to download it.
A new option "--nodownload" is added to that effect.

Also refactor the logic for reloading the database out of the controller and into the more appropriate model class.