Wrap database update in transactions
In case of network issues to the database server,
we want to avoid inconsistent state of the database,
so we wrap the updates for a given dataset in transactions.
That way, either everything or nothing related to that dataset are
changed in the DB.
If there is a network failure, we don't have to worry wether a dataset was partially updated.