Ask user to confirm prior executing command if datasets would be changed
When running the command in real mode (not dry run mode), the command
will ask the user whether he really want to proceed or not if 
the command would result in datasets to be changed in the database.

Also show user helpful message: confirmation of command execution, or
indicating the command is aborted. 
And if there is no more dataset left that need url replacement, a message
will be displayed to the user as well.


In addition of specific functional tests to verify the above behaviour, the acceptance test
scenario steps have been updated to handle the confirmation prompt non-interactively.

Finally, updated the acceptance tests hook so that the database is reloaded only when necessary
instead of before each scenario.