Add confirmation prompt to the DB restore command
like for the replacement commnad, we need to ask the user of the
command that download and restore production backup to confirm they want to proceed.
As DB restoration is destructive.

Also for both commands, they now return ExitCode::NOPERM when the user do not confirm,
which make more sense.