Fix functional tests not using the test database
Introduce ./yii_test runner as alternative to ./yii when running Yii2 commands in functional tests.
That way even when calling them using ``system()``, the commands will use the test database.
To do so, has to fix the ``config/test.php`` config to mirror ``config/console.php``.

Add grouping to each functional test so we can distinctly run all the functional test for one or the other of the two commands.

Fix the output of the --config option to also shows the actual DB connection details Yii2 is using.