Harden functional tests suite and update documentation
When running the project two main commands in functional tests, we should use
the yii_test runner and not the yii runner.

To make sure we don't accidentally use yii runner the setup hook will verify
that it is not called in the test classes.

Also updated the documentation to explain the yii2 runners and clarify the instuctions
for configuring GigaDB app to use our legacy database container service from the tool's directory.