Write drop/restore output to a log file and update documentation further
instead of showing long lines of psql and pg_restore diagnostics to STDOUT and STDERR respectively,
they are written to a log files in the tool's directory ``drop_restore.log``.

Updated the documentation to expect acceptance tests to take a long time to run (which may appear stuck
now that the PostgreSQL diagnosis lines are not showing in the terminal).