Move reference data generation from build to deploy stage
because access to the target database is required, the feeds needs to
be created on the target host during deploy stage.

That means the file location for the feeds is not longered baked in the container image,
se we need to a new volume to bind mounted to the container service on the target host
to hold the feeds of reference data.