Enable workaround for when the latest backup is corrupted
If for some reason the latest backup is corrupted, the functional and acceptance test suites
won't be passing and database errors will be thrown.

At the same time, we need to be aware when this incident happens so we can inform the DB admins while
being able to workaround it so we are not blocked.

So I've updated the functional tests DatasetFilesCest to not try to download the latest backup in the last scenario that was still doing it.
(Added additional functional test scenarios and assertions to make sure we are not losing test coverage)

Now all the tests must assume that the backup file for the day before current day is already in the filesystem
and appropriately named.

Then if that file is corrupted, as the workaround,  we can copy another working backup file into a new file named after the latest backup.

To diagnostic a problem with a particular backup file, we might want to download a series of backup files from various days.
To help with that, I've added a ``--norestore`` option to the downloader tool so that we can just download the backup files without
loading them into the local database each time.

Finally, I've updated the README with a new section at the end to explain the workaround.