Fix a data loss situation due to a missing existence check
When the downloader is called for a date
for which there is no backup file, the downloader will still drop the configured database.

The change adds a check to make sure the backup file exists first before doing anything
to the database.

also removed an occurence I missed in previous commit of a call to the downloader with --latest option and not --nodownload
in the functional tests that caused the local database to be empty if latest backup file is corrupted.