Update coscmd config to match other components' approach
Configuration file for `coscmd` should go in 
`gigadb/app/tools/dataset-backup-tool/config` like all other 
config files (rclone, Yii2).

The .dist templates as well, because dataset-backup-tool is
a standalone tool deployed separately from Gigadb web app,
so the files that made its codebase and configuration should 
all be colocated within the tool directory.

When used on backup server, `coscmd` is used without 
having to specify path to the config, so we should aim to 
replicate that behaviour in the smoke tests. Also it improves 
readability of the code and reduce opportunity to typos.

The default path where `coscmd` is looking for its
configuration when no config path is specified on command 
line is `~/.cos.conf`, so I've added a bind-mounted volume in 
`docker-compose.yml` to enable that default path.

README has been updated to reflect the changes made to
coscmd config.