Fix public download urls for dataset files to be configurable
the public dowload url for dataset files was fixed to the one currently used on live site.

However we need the download url to function on non-live environment for end-to-end testing.

So the download url needs to be configurable, hence the introduction of the env variable FILES_PUBLIC_URL.

Also fixed the acceptance tests that was sharing scenario implementation with test for reviewer download url.
Now ReviewerSteps and CuratorSteps each have their own test step for testing reviewer's mockup download url
and public download urls respectively.