Fix automated tests for files-url-updater after production data is fixed
we cannot just download production backup to seed the automated functional and acceptance tests for the tool
because the production is now fixed by the tool, so to test the funcitonality we need
to take an old production backup that has unfixed url and make it a stable test data fixture.
At the same time, we need to make sure that the download and restore tool can still download production backup.

We do so by redacting all user personal info from the old database file thenexport it as binary dump.
We also prepare a specific restoration index list for that backup (``default_restore.list``)
and renamed the original ``pg_restore.list`` to a more useful name ``production8_1_restore.list``.

The backend code in ``DatasetFiles.php`` for loading a binary dump into the configured database needs to detect whether it is to load the default test data
or a production backup.

The controller ``DatasetFilesController.php``, is updated to add a ``--default`` option that signals intent to work with the test data file rather than the production backup files. That work includes being able to instruct the download tool to download the test data from a different remote location.

I've updated the ``README.md`` file to mention the --default workflow.

This work is implementing task gigascience/gigadb-website#751 for Story gigascience/gigadb-website#570