(Docs): Add documentation about GigaDB EC2 permissions policy
Documentation is provided about the policy for EC2 that enforces how it 
can be used by the Developers users group. Documentation can be found 
in the awsdocs directory. Information is also provided for installing the
AWS command line tool for decoding any unauthorised error messages
when trying to use EC2 service.