Enable RDS instance to be restored from database snapshots and backups (#204)
# Pull request for issue: #732

This is a pull request for the following functionalities:

* Ability to restore a database snapshot and automated backup into a new RDS instance.

## Changes to the provisioning

When executing` tf_init.sh`, the full path to a backup file created by the `files-url-updater` tool is required. Here is an example prompt and required response:
```
You need to specify a backup file created by the files-url-updater tool: ../../../../gigadb/app/tools/files-url-updater/sql/gigadbv3_20210929_v9.3.25.backup
```

### Creation of snapshots and backups

`rds-instance.tf` has been updated by setting `skip_final_snapshot=false` and providing an expression to create a value for the `final_snapshot_identifier` variable to create a final database snapshot on termination of a RDS instance. 

In addition, `rds-instance.tf` has been updated to automatically create backups between 3-6 am UTC by setting `backup_retention_period=5` days. Because `delete_automated_backups=false`,  automated backups are retained when the RDS instance is deleted.

### Restoration of snapshots and backups

Snapshots can be restored in a new RDS instance by providing a value for `snapshot_identifier` which is null by default, e.g. `snapshot-for-testing` available in the Hong Kong AWS region, on the command-line as follows:
```
$ terraform plan -var snapshot_identifier="snapshot-for-testing"
$ terraform apply -var snapshot_identifier="snapshot-for-testing"
$ terraform refresh
```

Automated backups can either be restored to its latest restorable time or to a specific time. Both require a `restore_to_point_in_time ` configuration block which, in `terraform.tf`, is `null` by default. It is overridden by the presence of an `override.tf` file which contains an actual `restore_to_point_in_time` configuration block:
```
module "rds" {
   restore_to_point_in_time = {
     source_dbi_resource_id = var.source_dbi_resource_id
     restore_time = var.utc_restore_time
     use_latest_restorable_time = var.use_latest_restorable_time
   }
 }
```

This `override.tf` is only copied into the environment directory if a user wants to restore an automated backup by using a `--restore-backup` command line flag:
```
# Without --restore-backup flag, override.tf is not copied into the environment directory
$ ../../../scripts/tf_init.sh --project gigascience/forks/pli888-gigadb-website --env staging --restore-backup
```

Terraform will automatically detect the presence of `override.tf` in the directory and add it to the provisioning process when creating the RDS instance.

The following command line variables are then used to configure the `restore_to_point_in_time` code block. For example, to restore a backup to its latest restorable time:
```
$ terraform apply -var source_dbi_resource_id="db-6GQU4LWFBZI34AOR5BW2MEQFLU" -var gigadb_db_database="" -var use_latest_restorable_time="true"
```
N.B. `gigadb_db_database` needs to be empty as it will be provided by the backup.

To use the backup to restore to a specific time:
```
$ terraform apply -var source_dbi_resource_id="db-6GQU4LWFBZI34AOR5BW2MEQFLU" -var gigadb_db_database="" -var utc_restore_time="2021-10-27T06:02:12+00:00"
```

N.B. `source_dbi_resource_id` is the identifier for the backup you want to restore and these identifiers can be listed by executing:
```
$ aws rds describe-db-instance-automated-backups
```

## Changes to the documentation

* `SETUP_CI_CD_PIPELINE.md` has been updated with the steps to perform an RDS restoration using snapshots and backups.
* `policy-ec2.md` and `policy-rds.md` have been updated with the latest version of the policies currently used in the GigaDB AWS infrastructure.

---

## Procedure for restoring database snapshot and backups

We assume there is an existing RDS instance that we want to terminate and replace with a new RDS instance containing data that has been restored from a PostgreSQL snapshot or backup.

### Restoration of database snapshots
```
# Go to environment directory
$ cd <path>/gigadb-website/ops/infrastructure/envs/staging

# Terminate existing RDS service
$ terraform destroy --target module.rds

# Restore database snapshot in Hong Kong AWS region
$ terraform plan -var snapshot_identifier="snapshot-for-testing"
$ terraform apply -var snapshot_identifier="snapshot-for-testing"
$ terraform refresh
```

Now check the endpoint of your new RDS instance as shown in the AWS console is the same as that already defined for your Gitlab  `gigadb_db_host` variable.

Dataset 100043 will have a title that is specific to this snapshot - https://your-staging-gigadb/dataset/100043

### Restoration of database backups
```
# Go to environment directory
$ cd <path>/gigadb-website/ops/infrastructure/envs/staging

# Terminate existing RDS service
$ terraform destroy --target module.rds

# Copy override.tf to staging environment by using --restore-backup flag
$ ../../../scripts/tf_init.sh --project gigascience/forks/pli888-gigadb-website --env staging --restore-backup

# Backups can either be restored to its latest restorable time or to a specific
# time. Test the following in AWS Hong Kong region

# To restore to latest restorable time - need to override database name as this 
# will come from the backup
$ terraform apply -var source_dbi_resource_id="db-6GQU4LWFBZI34AOR5BW2MEQFLU" -var gigadb_db_database="" -var use_latest_restorable_time="true"

# To restore to specific time in backup - need to override database name as this 
# will come from the backup
$ terraform apply -var source_dbi_resource_id="db-6GQU4LWFBZI34AOR5BW2MEQFLU" -var gigadb_db_database="" -var utc_restore_time="2021-10-27T06:02:12+00:00"
```

Now check the endpoint of your new RDS instance as shown in the AWS console is the same as that already defined for your Gitlab  `gigadb_db_host` variable.

Dataset 100016 will have a title that is specific to this backup - https://your-staging-gigadb/dataset/100016